agent_using_rich/requirements.txt
agno
rich
python-dotenv
google-generativeai